requirements.txt
async-timeout==5.0.1
bleak==2.0.0
certifi==2025.11.12
influxdb-client==1.49.0
minimalmodbus==2.1.1
paho-mqtt==2.1.0
pyserial==3.5
python-dateutil==2.9.0.post0
reactivex==4.1.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.2
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Devices.Radios==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
